GSL spherical random vector
float -> double

+++ photon/GLWindow.cpp
@@ -268,7 +268,7 @@ bool GLWindow::CreateGLWindow( LPCSTR title, int width, int height, int bits)
 }
 
 // Здесь будет происходить вся прорисовка
-int GLWindow::DrawGLScene(std::vector<float> &Matrix)   
+int GLWindow::DrawGLScene(std::vector<double> &Matrix)   
 {
 	glEnable(GL_DEPTH_TEST);
 	glClear( GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT );		// Очистить экран и буфер глубины
